remove fluid.contrib.layers.ctr_metric_bundle
