+++ python/paddle/fluid/contrib/layers/__init__.py
@@ -16,10 +16,7 @@ from . import nn
 from .nn import *
 
 from .rnn_impl import *
-from . import metric_op
-from .metric_op import *
 
 __all__ = []
 __all__ += nn.__all__
 __all__ += rnn_impl.__all__
-__all__ += metric_op.__all__
